+++ src/finlang/ffiAttribute.cs
@@ -8,7 +8,7 @@ namespace finlang;
 /// Foreign Function Interface. This attribute is used to mark a method as a foreign function.
 /// Classes marked with this attribute will NOT be transpiled to C99.
 /// </summary>
-[AttributeUsage(AttributeTargets.Class)]
+[AttributeUsage(AttributeTargets.Class | AttributeTargets.Method)]
 public class ffiAttribute : Attribute {
 
 }